+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - **New:** Added `displayHtml` option.
 - **New:** Added `html` type of `getSelects` method.
 - **New:** Added level support for single group.
+- **New:** Added `filterByDataLength` option.
 - **New:** Added `customFilter` option support.
 - **New:** Added optgroup `styler` support and updated parameter.
 - **New:** Added `check` and `uncheck` methods.